requirements.txt
python-dotenv~=1.1.1
fastapi[all]~=0.116.1
uvicorn~=0.35.0
pipecat-ai[google,silero,webrtc,deepgram]

loguru~=0.7.3
langchain_openai~=0.3.27
langchain~=0.3.25
